Rhombus problems with bug fixes (also added back in the other quadrilateral problems for parallelograms. those were lost in a previous commit).
The two rhombus problems are Page172Problem19 and Page170ClassroomExercise02
Bug fixes:
1. Bug: diagonal intersection properties not set for rhombus
Fix: When constructed from a quad, the rhombus diagonal properties are set to match those of the quad
Temporary Fix: When constructed from a set of segments, there are optional parameters to pass the validity of the diagonals and the intersection point

2. Bug: Implied Component Calculator not detecting polygons or segments formed by un-named intersections
Fix: CalculateAllImpliedPolygons() was moved to occur AFTER finding all implied Segment-Segment intersection points and adding those points to the complete list of points. Added the function HandleSegmentsAtIntersection() to create the new segments and inMiddles formed by the detected intersection. FindAllImpliedExtendedSegmentSegmentPoints() will call this function whenever an implied intersection point is discovered.

3. Bug: StructurallyEqual for Parallelogram does not catch difference if the passed object is of a type that is derived from the type of the current object. In effect, this meant that clauses such as Strengthened(Quadrilateral(...), Parallelogram(...)) were found to be structurally equal to clauses such as Strengthened(Quadrilateral(...), Rhombus(...))
Fix: In the StructurallyEqual() function for Parallelogram, used 'Object.ReferenceEquals(this.GetType(), obj.GetType())' to check whether the passed object's type exactly matches the type of the current object